requirements.txt
faiss-cpu==1.8.0.post1
pandas>=2.2.0
pyarrow>=15.0.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
transformers>=4.41.0
accelerate>=0.28.0
spacy>=3.7.0
langchain>=0.2.8
langchain-community>=0.2.7
ollama>=0.3.0
streamlit>=1.35.0
matplotlib>=3.8.0
